WIP
Rebase on gatling/main (thanks to @Isammoc)
