+++ src/docs/content/_index.md
@@ -20,7 +20,10 @@ Out of the box, Gatling comes with excellent support of the HTTP protocol that m
 As the core engine is actually protocol agnostic, it is perfectly possible to implement support for other protocols.
 For example, Gatling currently also ships JMS support.
 
-The [quickstart]({{< ref "tutorials/quickstart" >}}) has an overview of the most important concepts, walking you through the setup of a simple scenario for load testing an HTTP server.
+The following tutorials will help you get started with Gatling:  
+[Introduction to scripting]({{< ref "tutorials/scripting-intro/index" >}})  
+[Introduction to the Recorder]({{< ref "tutorials/recorder-tutorial/index" >}})  
+[Writing DRY tests]({{< ref "tutorials/advanced/index" >}})  
 
 Having *scenarios* that are defined in code and are resource efficient are the two requirements that motivated us to create Gatling. Based on an expressive [DSL](http://en.wikipedia.org/wiki/Domain-specific_language), the *scenarios* are self-explanatory. They are easy to maintain and can be kept in a version control system.
