+++ src/docs/content/tutorials/advanced/index.md
@@ -1,5 +1,5 @@
 ---
-title: "Advanced Tutorial"
+title: "Writing DRY tests"
 description: "How to transform a raw recorded test into a dynamic and meaningful load test. Introduce feeders, dynamic parameters, loops, functions and multiple scenarios."
 lead: "Isolate process, configure virtual users, use feeders, checks and looping"
 date: 2021-04-20T18:30:56+02:00
@@ -7,7 +7,7 @@ lastmod: 2022-12-14T21:30:56+02:00
 weight: 1030000
 ---
 
-In this section, we assume that you have already gone through the [Quickstart]({{< ref "../quickstart" >}}) section and that you have a basic simulation to work with.
+In this section, we assume that you have already gone through the [Introduction to the Recorder](link) section and that you have a basic simulation to work with.
 We will apply a series of refactorings to introduce more advanced concepts and [Domain Specific Language](https://en.wikipedia.org/wiki/Domain-specific_language) constructs.
 
 {{< alert tip >}}
