+++ src/docs/content/tutorials/recorder-tutorial/code/BasicSimulationKotlin.kt
