+++ src/docs/content/tutorials/recorder-tutorial/images/recorder.png
