+++ src/docs/content/tutorials/scripting-intro/includes/preparation.java.md
@@ -1,0 +1,1 @@
+test
