+++ src/docs/content/tutorials/scripting-intro/includes/preparation.kotlin.md
@@ -1,0 +1,1 @@
+test
